# Small optimizations for decoders