This is a quick attempt to gather some of the low-hanging fruit I noticed in #343.

The issue is that `Xor.fromOption` creates three unnecessary function objects, which isn't ideal in situations like `decodeInt`, where it's going to get called for every JSON value we're trying to decode.

The wins aren't huge, but they're not negligible, and there's not a big difference in code readability.

Before:

```
DecodingBenchmark.decodeFoosC  thrpt   80   3379.073 ± 16.250  ops/s
DecodingBenchmark.decodeIntsC  thrpt   80  19272.072 ± 96.131  ops/s

DecodingBenchmark.decodeFoosC:gc.alloc.rate.norm  thrpt    5  1185168.346 ±     0.007    B/op
DecodingBenchmark.decodeIntsC:gc.alloc.rate.norm  thrpt    5   275349.679 ±    48.383    B/op
```

After:

```
DecodingBenchmark.decodeFoosC  thrpt   80   3383.712 ± 10.645  ops/s
DecodingBenchmark.decodeIntsC  thrpt   80  20376.707 ± 63.134  ops/s

DecodingBenchmark.decodeFoosC:gc.alloc.rate.norm  thrpt    5  1151528.338 ±     0.022    B/op
DecodingBenchmark.decodeIntsC:gc.alloc.rate.norm  thrpt    5   257284.347 ±    36.937    B/op
```